Dev-2.0.0 modify configuraiton loading
